+++ apiserver/service/src/main/java/com/zipline/service/message/MessageTemplateService.java
@@ -1,6 +1,6 @@
 package com.zipline.service.message;
 
-import com.zipline.service.message.dto.message.request.MessageTemplateRequestDTO;
+import com.zipline.service.message.dto.request.MessageTemplateRequestDTO;
 
 public interface MessageTemplateService {
   void createMessageTemplate(MessageTemplateRequestDTO requestDTO, Long userUid);
